feat: delete ArchivedComponents folder
