+++ sonar-project.properties
@@ -71,7 +71,6 @@ sonar.coverage.exclusions=\
   src/node-lib/**/*.fixture.ts, \
   src/utils/curriculum/**/*.fixture.ts, \
   src/utils/gCloudStorage.ts, \
-  src/components/ArchivedComponents/**, \
   src/pages/pupils/beta/ab-demos/**, \
   src/node-lib/posthog/**, \
   src/app/(registration)/**, \
@@ -109,7 +108,6 @@ sonar.exclusions=\
   src/node-lib/sanity-graphql/generated/sdk.ts,\
   src/browser-lib/cookie-consent/confirmic/*, \
   src/browser-lib/avo/Avo.ts, \
-  src/components/ArchivedComponents/**, \
   # exclude json files (https://community.sonarsource.com/t/json-file-error-expected-an-assignment-or-function-call-and-instead-saw-an-expression/96301/13)
   **/*.json
