--- src/components/ArchivedComponents/BrowserWidthBar/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./BrowserWidthBar";
