--- src/components/ArchivedComponents/DynamicForm/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./DynamicForm";
