--- src/components/ArchivedComponents/ExpandingContainer/index.tsx
@@ -1,1 +1,0 @@
-export { default } from "./ExpandingContainer";
