--- src/components/ArchivedComponents/Home/index.ts
@@ -1,2 +1,0 @@
-export { default as SharedHomeContent } from "./SharedHomeContent";
-export { default as HomeSiteCards } from "./HomeSiteCards";
