--- src/components/ArchivedComponents/IconBackground/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./IconBackground";
