--- src/components/ArchivedComponents/ImageBox/index.tsx
@@ -1,2 +1,0 @@
-export { default } from "./ImageBox";
-export type { ImageBoxProps } from "./ImageBox";
