--- src/components/ArchivedComponents/KeyStagesNav/KeyStagesNav.stories.tsx
@@ -1,18 +1,0 @@
-import { StoryFn, Meta } from "@storybook/react";
-
-import Component from "./KeyStagesNav";
-
-import keyStagesNavData from "@/browser-lib/fixtures/keyStagesNav";
-
-export default {
-  component: Component,
-  argTypes: {},
-} as Meta<typeof Component>;
-
-const Template: StoryFn<typeof Component> = () => {
-  return <Component keyStages={keyStagesNavData} />;
-};
-
-export const KeyStagesNav = {
-  render: Template,
-};
