--- src/components/ArchivedComponents/KeyStagesNav/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./KeyStagesNav";
