--- src/components/ArchivedComponents/LoadingSpinner/LoadingSpinner.stories.tsx
@@ -1,9 +1,0 @@
-import { Meta } from "@storybook/react";
-
-import Component from "./LoadingSpinner";
-
-export default {
-  component: Component,
-} as Meta<typeof Component>;
-
-export const LoadingSpinner = {};
