--- src/components/ArchivedComponents/LoadingSpinner/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./LoadingSpinner";
