--- src/components/ArchivedComponents/Toast/Toast.test.tsx
@@ -1,37 +1,0 @@
-import userEvent from "@testing-library/user-event";
-
-import Toast from "./Toast";
-
-import ToastProvider from "@/context/Toast/ToastProvider";
-import renderWithTheme from "@/__tests__/__helpers__/renderWithTheme";
-import waitForNextTick from "@/__tests__/__helpers__/waitForNextTick";
-import CopyLinkButton from "@/components/SharedComponents/Button/CopyLinkButton";
-
-describe("toast notification", () => {
-  const ROLE = "alert";
-  test("shows feedback", async () => {
-    const { getByRole, getByTitle, rerender } = renderWithTheme(
-      <ToastProvider>
-        <CopyLinkButton />
-        <Toast />
-      </ToastProvider>,
-    );
-
-    const user = userEvent.setup();
-
-    const button = getByTitle("Copy link to clipboard");
-    expect(button).toBeInTheDocument();
-
-    await user.click(button);
-
-    await waitForNextTick();
-
-    rerender(
-      <ToastProvider>
-        <Toast />
-      </ToastProvider>,
-    );
-
-    expect(getByRole(ROLE)).toHaveTextContent("Link copied to clipboard");
-  });
-});
