--- src/components/ArchivedComponents/Toggle/index.tsx
@@ -1,1 +1,0 @@
-export { default } from "./Toggle";
