+++ src/components/introduction.mdx
@@ -9,7 +9,6 @@ Welcome to Oak's Storybook. Browse our component examples and usage guidelines. 
 
 ## Folder structure:
 - AppComponents
-- ArchivedComponents
 - CurriculumComponents
 - GenericPagesComponents
 - GenericPagesViews
@@ -24,8 +23,6 @@ Welcome to Oak's Storybook. Browse our component examples and usage guidelines. 
 
 - **AppComponents** 
     - app related components such as header, footer, menu, layout, etc.
-- **ArchivedComponents**
-    - components that are not imported aynwhere
 - **CurriculumComponents**
     - curriculum related components (components used in `/teachers/curriculum/` belong here rather than in `TeacherComponents`)
 - **GenericPagesComponents**
